refactor: colocate @react-email/button source in react-email (#3233)
