--- packages/button/CHANGELOG.md
@@ -1,94 +1,0 @@
-# @react-email/button
-
-## 0.2.1
-
-### Patch Changes
-
-- 8b7a660: remove use of devEngines which npm detects
-
-## 0.2.0
-
-### Minor Changes
-
-- e2fa065: respect the order in which padding properties are defined
-
-## 0.1.0
-
-### Minor Changes
-
-- 4c4c5ef: respect the order in which padding properties are defined
-
-## 0.1.0-canary.0
-
-### Minor Changes
-
-- 11c4600: respect the order in which padding properties are defined
-
-## 0.0.19
-
-### Patch Changes
-
-- 467af4e: Use peerDependencies for react and react-dom
-- b34aa90: Move react and react-dom to just dependencies for better DX
-
-## 0.0.19-canary.1
-
-### Patch Changes
-
-- 467af4e: Use peerDependencies for react and react-dom
-
-## 0.0.19-canary.0
-
-### Patch Changes
-
-- f7833da: Move react and react-dom to just dependencies for better DX
-
-## 0.0.18
-
-### Patch Changes
-
-- 918e62c: Fix NaN values on rendered Button HTML
-
-## 0.0.18-canary.0
-
-### Patch Changes
-
-- bd4747f: Fix NaN values on rendered Button HTML
-
-## 0.0.17
-
-### Patch Changes
-
-- 3caaf53: Updated peer dependencies to allow for React 19 release candidated and React 19 itself
-- bec4b08: Add missing `msoPaddingAlt` to containing `<a>` tag
-
-## 0.0.17-canary.1
-
-### Patch Changes
-
-- 8538cb3: Add missing `msoPaddingAlt` to containing `<a>` tag
-
-## 0.0.17-canary.0
-
-### Patch Changes
-
-- a1c016b: Updated peer dependencies to allow for React 19 release candidated and React 19 itself
-
-## 0.0.16
-
-### Patch changes
-
-- Fixes horizontal padding not working on Outlook for Windows
-- Add forward ref
-
-## 0.0.16-canary.1
-
-### Patch changes
-
-- Fixes horizontal padding not working on Outlook for Windows
-
-## 0.0.16-canary.0
-
-### Patch changes
-
-- Add forward ref
