--- packages/button/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@react-email/button",
-  "version": "0.2.1",
-  "description": "A link that is styled to look like a button",
-  "sideEffects": false,
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.cts",
-  "files": [
-    "dist/**"
-  ],
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.mts",
-        "default": "./dist/index.mjs"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    }
-  },
-  "license": "MIT",
-  "scripts": {
-    "build": "tsdown src/index.ts --format esm,cjs --dts --external react",
-    "build:watch": "tsdown src/index.ts --format esm,cjs --dts --external react --watch",
-    "clean": "rm -rf dist",
-    "test": "vitest run",
-    "test:watch": "vitest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/resend/react-email.git",
-    "directory": "packages/button"
-  },
-  "keywords": [
-    "react",
-    "email"
-  ],
-  "engines": {
-    "node": ">=20.0.0"
-  },
-  "peerDependencies": {
-    "react": "^18.0 || ^19.0 || ^19.0.0-rc"
-  },
-  "devDependencies": {
-    "@react-email/render": "workspace:*",
-    "tsconfig": "workspace:*",
-    "typescript": "catalog:"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
