+++ packages/react-email/src/components/button/button.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { parsePadding } from './utils/parse-padding';
-import { pxToPt } from './utils/px-to-pt';
+import { parsePadding } from './utils/parse-padding.js';
+import { pxToPt } from './utils/px-to-pt.js';
 
 export type ButtonProps = Readonly<React.ComponentPropsWithoutRef<'a'>>;
