+++ packages/react-email/src/components/button/index.ts
@@ -1,0 +1,1 @@
+export * from './button.js';
