+++ packages/react-email/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './body/index.js';
+export * from './button/index.js';
