+++ packages/react-email/tsconfig.json
@@ -34,6 +34,6 @@
     "types": ["vitest/globals"],
     "outDir": "dist"
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "src/**/*.tsx"],
   "exclude": ["dist", "node_modules"]
 }
